requirements.txt
asgiref==3.4.1
autopep8==1.6.0
click==8.0.4
colorama==0.4.4
Django==3.2.12
django-ckeditor==6.3.2
django-crispy-forms==1.13.0
django-js-asset==2.0.0
django-online-users==0.3
django-tinymce==3.4.0
djlint==0.7.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
pathspec==0.9.0
Pillow==8.4.0
pycodestyle==2.8.0
python-decouple==3.6
pytz==2021.3
PyYAML==6.0
regex==2022.3.15
sqlparse==0.4.2
toml==0.10.2
tomlkit==0.8.0
tqdm==4.64.0
typing_extensions==4.0.1
zipp==3.6.0
